telegram-bot/requirements.txt
gunicorn
flask
python-telegram-bot